backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
Pillow
exifread
folium
python-multipart 
segmentation-models-pytorch 